requirements.txt
streamlit
os
json
google-generativeai
python-dotenv
